docs: architecture and changelog, feat: modular frontend structure, and refactor: backend performance optimizations

+++ td-backend/src/main/java/com/lichiapps/techdepot/dtos/ItemDTO.java
@@ -27,9 +27,9 @@ public class ItemDTO {
     @Data
     public static class EspecificacionesDTO {
         private Integer largo;
-        private Integer amperajeMax;
-        private Integer amperaje;
-        private Integer voltaje;
+        private Short amperajeMax;
+        private Short amperaje;
+        private Short voltaje;
         private String modelo;
     }
